refactor : CommaPatternNames에서 이름과 분리라는 두가지 책임 분리
