+++ src/main/java/racinggame/RacingMain.java
@@ -1,11 +1,8 @@
 package racinggame;
 
 import racinggame.controller.GameController;
-import racinggame.domain.CommaPatternNames;
 import racinggame.dto.Result;
-import racinggame.random.DefaultRandomNumberGenerator;
 import racinggame.view.OutPutView;
-import racinggame.domain.RacingGame;
 import racinggame.dto.InputDto;
 import racinggame.view.InputView;
 
@@ -15,7 +12,7 @@ public class RacingMain {
         String carNames = InputView.readCarNames();
         int repeatCount = InputView.readRepeatCount();
         GameController controller = new GameController();
-        Result result = controller.play(new InputDto(carNames,repeatCount));
+        Result result = controller.play(new InputDto(carNames, repeatCount));
         OutPutView.render(result);
     }
