+++ src/main/java/racinggame/controller/GameController.java
@@ -1,6 +1,7 @@
 package racinggame.controller;
 
-import racinggame.domain.CommaPatternNames;
+import racinggame.domain.CommaSplitter;
+import racinggame.domain.Names;
 import racinggame.domain.RacingGame;
 import racinggame.dto.InputDto;
 import racinggame.dto.Result;
@@ -10,7 +11,7 @@ public class GameController {
 
     public Result play(InputDto inputDto) {
         RacingGame racingGame = new RacingGame(inputDto.getRepeatCount(),
-                new CommaPatternNames(inputDto.getCarNames()));
+                new Names(inputDto.getCarNames(), new CommaSplitter()));
         return racingGame.start(new DefaultRandomNumberGenerator());
     }
 }
