+++ src/main/java/racinggame/domain/Names.java
@@ -1,8 +1,17 @@
 package racinggame.domain;
 
+import java.util.Collections;
 import java.util.List;
 
-public interface Names {
+public class Names {
 
-    List<String> split();
+    private final List<String> names;
+
+    public Names(String names, PatternSplitter patternSplitter) {
+        this.names = Collections.unmodifiableList(patternSplitter.split(names));
+    }
+
+    public List<String> split() {
+        return names;
+    }
 }
