+++ src/main/java/racinggame/domain/PatternSplitter.java
@@ -1,0 +1,8 @@
+package racinggame.domain;
+
+import java.util.List;
+
+public interface PatternSplitter {
+
+    List<String> split(String string);
+}
